Updated model and view In Serializers

+++ HRMS/GM.py
@@ -1,36 +1,49 @@
 from cgitb import lookup
 from rest_framework.response import Response
-from rest_framework import   status
+from rest_framework import status
 from rest_framework.views import APIView
 
 
-
-class GenericMethods():
-    def getall(Model , ModelSerializer):
+class GenericMethods:
+    def getall(Model, ModelSerializer):
         try:
-            return Response(ModelSerializer(Model.objects.all(),many = True).data,status=status.HTTP_200_OK)
-        except :
-            return Response(data = {'Error' : str(Model._meta).split('.')[1]+' object does not exists'},status=status.HTTP_400_BAD_REQUEST)
+            return Response(
+                ModelSerializer(Model.objects.all(), many=True).data,
+                status=status.HTTP_200_OK,
+            )
+        except:
+            return Response(
+                data={
+                    "Error": str(Model._meta).split(".")[1] + " object does not exists"
+                },
+                status=status.HTTP_400_BAD_REQUEST,
+            )
 
-    def getone(Model,ModelSerializer,pk):
+    def getone(Model, ModelSerializer, pk):
         try:
-            return Response(ModelSerializer(Model.objects.get(id=pk)).data,status=status.HTTP_200_OK)
+            return Response(
+                ModelSerializer(Model.objects.get(id=pk)).data,
+                status=status.HTTP_200_OK,
+            )
         except Model.DoesNotExist:
-            return Response(data = {'Error' : str(Model._meta).split('.')[1]+' object does not exists'},status=status.HTTP_400_BAD_REQUEST)
+            return Response(
+                data={
+                    "Error": str(Model._meta).split(".")[1] + " object does not exists"
+                },
+                status=status.HTTP_400_BAD_REQUEST,
+            )
 
-
-    def post(ModelSerializer,data):
+    def post(ModelSerializer, data):
         serializer = ModelSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
-            return Response(data = serializer.data , status=status.HTTP_201_CREATED)
+            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(data = serializer.errors,status=status.HTTP_200_OK)
+            return Response(data=serializer.errors, status=status.HTTP_200_OK)
 
-
-    def put(Model,ModelSerializer,data,id):
-        classroom=Model.objects.get(id=id)
-        serializer=ModelSerializer(classroom, data=data)
+    def put(Model, ModelSerializer, data, id):
+        classroom = Model.objects.get(id=id)
+        serializer = ModelSerializer(classroom, data=data)
         if serializer.is_valid():
             serializer.save()
             return Response(data=serializer.data, status=status.HTTP_202_ACCEPTED)
@@ -38,10 +51,9 @@ class GenericMethods():
 
 
 class GenericMethodsMixin:
-    
     def __init__(self) -> None:
         self.model = self.get_model()
-        self.queryset=self.get_queryset()
+        self.queryset = self.get_queryset()
         self.serializer = self.get_serializer_class()
         self.lookup = self.get_lookup()
 
@@ -56,40 +68,77 @@ class GenericMethodsMixin:
 
     def get_queryset(self):
         return self.model.objects.all()
-    
-    def get(self,request,pk,*args, **kwargs):
-        filter = {self.lookup:pk}
-        if pk ==0:
+
+    def get(self, request, pk, *args, **kwargs):
+        filter = {self.lookup: pk}
+        if pk == 0:
             try:
-                return Response(self.serializer(self.model.objects.all(),many = True).data,status=status.HTTP_200_OK)
-            except :
-                return Response(data = {'Error' : str(self.model._meta).split('.')[1]+' object does not exists'},status=status.HTTP_400_BAD_REQUEST)
+                return Response(
+                    self.serializer(self.model.objects.all(), many=True).data,
+                    status=status.HTTP_200_OK,
+                )
+            except:
+                return Response(
+                    data={
+                        "Error": str(self.model._meta).split(".")[1]
+                        + " object does not exists"
+                    },
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
         else:
             try:
-                return Response(self.serializer(self.model.objects.get(pk=pk)).data,status=status.HTTP_200_OK)
+                return Response(
+                    self.serializer(self.model.objects.get(pk=pk)).data,
+                    status=status.HTTP_200_OK,
+                )
             except self.model.DoesNotExist:
-                return Response(data = {'Error' : str(self.model._meta).split('.')[1]+' object does not exists'},status=status.HTTP_400_BAD_REQUEST)
+                return Response(
+                    data={
+                        "Error": str(self.model._meta).split(".")[1]
+                        + " object does not exists"
+                    },
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
 
-    def post(self,request,*args, **kwargs):
+    def post(self, request, *args, **kwargs):
         serializer = self.serializer(data=request.data)
+        print(request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(data = serializer.data , status=status.HTTP_201_CREATED)
+            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(data = serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+            return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def put(self,request,pk,*args, **kwargs):
-        filter = {self.lookup:pk}
-        classroom=self.model.objects.get(**filter)
-        serializer=self.serializer(classroom, data=request.data)
+
+    def put(self, request, pk, *args, **kwargs):
+        filter = {self.lookup: pk}
+        object1 = self.model.objects.get(**filter)
+        if request.user != object1.created_by:
+            return Response(
+                {"message : Unauthorized"}, status=status.HTTP_401_UNAUTHORIZED
+            )
+        serializer = self.serializer(object1, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(data=serializer.data, status=status.HTTP_202_ACCEPTED)
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    def delete(self,request,pk,*args,**kwargs) :
+
+
+    def delete(self, request, pk, *args, **kwargs):
         data = self.model.objects.get(pk=pk)
-        if data : 
+        if request.user != data.created_by:
+            return Response(
+                {"message : Unauthorized"}, status=status.HTTP_401_UNAUTHORIZED
+            )
+        if data:
             data.delete()
-            return Response({"data" : "Record Deleted Successfully"},status=status.HTTP_204_NO_CONTENT)
-        return Response(data = {'Error' : str(self.model._meta).split('.')[1]+' object does not exists'},status=status.HTTP_400_BAD_REQUEST)
+            return Response(
+                {"data": "Record Deleted Successfully"},
+                status=status.HTTP_204_NO_CONTENT,
+            )
+        return Response(
+            data={
+                "Error": str(self.model._meta).split(".")[1] + " object does not exists"
+            },
+            status=status.HTTP_400_BAD_REQUEST,
+        )
